SNOW-2176180: rename to workload-identity from workflow-identity
