+++ src/snowflake/cli/_plugins/auth/__init__.py
@@ -1,6 +1,6 @@
 from snowflake.cli._plugins.auth.keypair.commands import app as keypair_app
-from snowflake.cli._plugins.auth.workflow_identity.commands import (
-    app as workflow_identity_app,
+from snowflake.cli._plugins.auth.workload_identity.commands import (
+    app as workload_identity_app,
 )
 from snowflake.cli.api.commands.snow_typer import SnowTyperFactory
 from snowflake.cli.api.feature_flags import FeatureFlag
@@ -12,4 +12,4 @@ app = SnowTyperFactory(
 )
 
 app.add_typer(keypair_app)
-app.add_typer(workflow_identity_app)
+app.add_typer(workload_identity_app)
