+++ src/snowflake/cli/_plugins/auth/workload_identity/__init__.py
