+++ src/snowflake/cli/_plugins/auth/workload_identity/oidc_providers.py
