+++ src/snowflake/cli/_plugins/auth/workload_identity/plugin_spec.py
